+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.1] - 2023-08-07
 
 ### Fixed
-- Incorrect location being reported for unterminated strings.
+- Incorrect location being reported for unterminated quoted strings.
 
 ## [0.2.0] - 2023-08-04