+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2023-08-08
 
 ### Fixed
 - Incorrect location being reported for unterminated C style comments.
@@ -28,7 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2023-08-02
 
-[unreleased]: https://github.com/IanE9/qentities/compare/v0.2.1...HEAD
+[unreleased]: https://github.com/IanE9/qentities/compare/v0.2.2...HEAD
+[0.2.2]: https://github.com/IanE9/qentities/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/IanE9/qentities/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/IanE9/qentities/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/IanE9/qentities/releases/tag/v0.1.0